+++ examples/read_buffered.rs
@@ -23,7 +23,7 @@ fn main() -> Result<(), quick_xml::Error> {
                 count += 1;
             }
             Ok(Event::Eof) => break, // exits the loop when reaching end of file
-            Err(e) => panic!("Error at position {}: {:?}", reader.buffer_position(), e),
+            Err(e) => panic!("Error at position {}: {:?}", reader.error_position(), e),
             _ => (), // There are several other `Event`s we do not consider here
         }
     }