+++ src/reader/buffered_reader.rs
@@ -8,7 +8,8 @@ use std::path::Path;
 use crate::errors::{Error, Result};
 use crate::events::Event;
 use crate::name::QName;
-use crate::reader::{BangType, Parser, ReadTextResult, Reader, Span, XmlSource};
+use crate::parser::Parser;
+use crate::reader::{BangType, ReadTextResult, Reader, Span, XmlSource};
 use crate::utils::is_whitespace;
 
 macro_rules! impl_buffered_source {
@@ -98,54 +99,6 @@ macro_rules! impl_buffered_source {
 
             *position += read;
             ReadTextResult::UpToEof(&buf[start..])
-        }
-
-        #[inline]
-        $($async)? fn read_bytes_until $(<$lf>)? (
-            &mut self,
-            byte: u8,
-            buf: &'b mut Vec<u8>,
-            position: &mut u64,
-        ) -> io::Result<(&'b [u8], bool)> {
-            // search byte must be within the ascii range
-            debug_assert!(byte.is_ascii());
-
-            let mut read = 0;
-            let start = buf.len();
-            loop {
-                let available = match self $(.$reader)? .fill_buf() $(.$await)? {
-                    Ok(n) if n.is_empty() => break,
-                    Ok(n) => n,
-                    Err(ref e) if e.kind() == io::ErrorKind::Interrupted => continue,
-                    Err(e) => {
-                        *position += read;
-                        return Err(e);
-                    }
-                };
-
-                match memchr::memchr(byte, available) {
-                    Some(i) => {
-                        buf.extend_from_slice(&available[..i]);
-
-                        let used = i + 1;
-                        self $(.$reader)? .consume(used);
-                        read += used as u64;
-
-                        *position += read;
-                        return Ok((&buf[start..], true));
-                    }
-                    None => {
-                        buf.extend_from_slice(available);
-
-                        let used = available.len();
-                        self $(.$reader)? .consume(used);
-                        read += used as u64;
-                    }
-                }
-            }
-
-            *position += read;
-            Ok((&buf[start..], false))
         }
 
         #[inline]
@@ -327,7 +280,7 @@ impl<R: BufRead> Reader<R> {
     ///     match reader.read_event_into(&mut buf) {
     ///         Ok(Event::Start(_)) => count += 1,
     ///         Ok(Event::Text(e)) => txt.push(e.unescape().unwrap().into_owned()),
-    ///         Err(e) => panic!("Error at position {}: {:?}", reader.buffer_position(), e),
+    ///         Err(e) => panic!("Error at position {}: {:?}", reader.error_position(), e),
     ///         Ok(Event::Eof) => break,
     ///         _ => (),
     ///     }